pass flag to not connect to network table

+++ RaspberryPi/main.py
@@ -4,6 +4,7 @@ import vision as vision
 
 from ntClient import *
 import os
+import sys
 import time
 
 # Id's for reading data from networktable
@@ -21,7 +22,8 @@ prefSideLeftId = "prefSideLeft"
 # if this script is run directly by python, then __name__ is '__main__'.  If it
 # is run because it is imported, then __name__ is the module name.
 
-if __name__ == '__main__': 
+if __name__ == '__main__':
+ 
 	logging.basicConfig(filename='/var/log/robovision/robovision.log',level=logging.DEBUG)
 	logging.debug("vision services started.")
 	#logging.debug("uid {} euid {}".format(os.getuid(), os.geteuid()))
@@ -30,28 +32,41 @@ if __name__ == '__main__':
 	tableName = "955"
 	tableDirectory = '/' + tableName + '/'
 
-	# test for table existing by setting a value.  If table not accessible, wait and try.
-	while True:
-		try:
-			table = NetworkTableClient(tableName)
-			logging.debug("Network table initialized.")
-			table.setValue(tableDirectory + isGoalHotId, False) # Set foundHotTarget
-			break # leave the while loop if the table is accessible
-		except Exception as ex:
-			logging.error(ex)
-			time.sleep(1/4.)
+	table = None
+	runVision = True
+	debugMode = True
+	prefSideLeft = True
+
+	# Test for table existing by setting a value.  If table not accessible, wait and try.
+	# Don't use network table if we pass "-n" option on command line.
+	if not (len(sys.argv) > 1 and sys.argv[1] == '-n'):
+		logging.debug("using network table.")
+		while True:
+			try:
+				logging.debug("initialize network table.")
+				table = NetworkTableClient(tableName)
+				logging.debug("Network table initialized.")
+				table.setValue(tableDirectory + isGoalHotId, False) # Set foundHotTarget
+				break # Leave the while loop if the table is accessible
+			except:
+				logging.error("error initializing network table.")
+				time.sleep(1/4.)
+	else:
+		logging.debug("not using network table.")
 
 	try:
 		while True:
-			runVision = table.getValue(tableDirectory + runVisionId)
-			debugMode = table.getValue(tableDirectory + debugId)
-			prefSideLeft = table.getValue(tableDirectory + prefSideLeftId)
+			if table:
+				runVision = table.getValue(tableDirectory + runVisionId)
+				debugMode = table.getValue(tableDirectory + debugId)
+				prefSideLeft = table.getValue(tableDirectory + prefSideLeftId)
 
 			if runVision: # Check if runVision is true
 				vision.update(prefSideLeft, debugMode)	# Run vision, check whether debug mode is true
-				table.setValue(tableDirectory + isGoalHotId, vision.getFoundHotTarget()) # Set foundHotTarget
+				if table:
+					table.setValue(tableDirectory + isGoalHotId, vision.getFoundHotTarget()) # Set foundHotTarget
 
-			if table.getValue(tableDirectory + shutDownPi):
+			if table and table.getValue(tableDirectory + shutDownPi):
 				os.system("sudo shutdown -h now")
 				break
